some more sanitized URLs
